Removed death animation from blend tree in wizard animation controller
Accidentally left behind an old implementation of the blend tree where the death state was not its own state but a part of the motion blend tree.

This would cause NPC wizards to accidentally play the last frame of the death animation from time to time while casting, leading to a funny but unintended situation where they would constantly fall to the ground, flat on their faces, lmao....

This has now been fixed.