Added ownership system to spell caster controller
This change allows the spell caster to spawn beams that do not damage the owner.

This change has been made because sometimes, beams have to spawn in positions which may slightly intersect with the hitbox of their owner, and obviously damaging the owner is not the correct behaviour.

This could also be used in the future to prevent projectiles from hitting the owner if the owner throws them in a given direction while moving at a at high speed toward the same direction, which would cause the owner to collide with their own projectile.