Cleaned up NPC Controller and removed FSM and US logic
Cleaned up the organization of the NPC Controller class.

The FSM and Utility System logic has been removed from this class, as it is going to be moved into the specific behaviour controller classes to make it easier to implement unique behaviours for each NPC.

This is yet another step toward making NPC Controller a generic class for NPC control, manipulation and action execution as well as a sensor system and information gathering point for other subsystems.