Fixed bug in sight sensor and implemented detection over time system
There was a bug in the sight sensor detection system that had gone unnoticed until now, where the sensor would simply not detect anything at all. This was caused by thefact that the player's origin point is on its feet, which would cause the detection line trace to always hit the ground first.

This has now been fixed by adding a small tolerance value and casting the ray higher.

The sight system has been modified to make use of the new detection over time system. The sight detection depends on the distance to the player.

The debug drawing system for sight detection has also been improved.