Modified AlertUI Controller to use OnValueChanged event for detection progress
With this new change, the code not only does not require updating all of the state every single frame, it also simplified the code a lot and made it easier to fix an issue where damaging enemies did not let the icon become visible at all, since on attack the value is set direction to 1.0f, which is 100% and that meant that the condition to set the canvas group's alpha value to 1.0f was never met. The NPC behaviour logic works correctly, the only problem was that the alert icon simply never became visible.