Modified NPC Death handling to stop casting
This change fixes an issue where dead NPCs that were casting a spell such as a beam would continue casting until the spell was finished or until they were completely destroyed.

In the case of beam spells, this would leave behind floating beam spells.

In the case of the other spells, this would allow NPCs to cast queued up spells after death during the period of time where the death animation is playing.

To solve this issue, the stop casting method is called.

TODO : If the NPC is queuing up in the Attack() coroutine a set of elements to start casting, the issue persists. A method must be added to the spell caster class to empty the element queue instantly.