Added command CmdLeaveMeAlone()
The command CmdLeaveMeAlone() fetches all of the NPCs in the scene and resets their state by setting their targets to null and their detection progress to 0, as well as calling the reset behaviour method on their behaviour script.

With this cheat command, the player can make all NPCs forget about the player and quite literally be left alone. Basically, an useful command for testing and debugging, and also a cool cheat.