Fixed flee script by increasing flee target point distance
Due to the fact that the flee vector was multiplied by a small value, the flee target point calculated was right next to the origin point of the fleeing NPC. This meant that the NPC would stand in place twitching slightly in the flee direction rather than properly fleeing.

This has now been fixed by increasing the distance used to calculate the flee-to target nav point.