Added isAttacking flag to NPCController
The NPC Controller script now has an isAttacking flag which prevents it from accidentally starting an attack while already attacking.

The lack of this flag used to cause a bug where sometimes a spell would start being casted while a beam spell was in the process of being casted through the start attack coroutine without calling stop casting, which would make it so that the spell caster would not keep the beam reference in memory and said beam would never despawn.

This has now been fixed.

TODO : Clean up the code cause it's looking pretty sloppy...