Implemented music volume level system
The sound level system allows setting volume by increments in the settings menu without having many problems with floating point accuracy issues.

This system is reminiscent of old games where volume settings would usually move in increments. Nowadays, most games use a continuous bar, but that gives a bad degree of accuracy for setting volume settings in my opinion, so this is a temporary implementation to get that system back.

In the future, make an increments system as well as a slider system support so that users can set it however they want, because having both systems is actually pretty useful. Also add a command to set the volume settings to some specific value.