Modified RoomController to use SetActive() rather than destroying walls
This change allows the dungeon to re-generate walls if need be in the future.

This change also makes it easier to generate a nav mesh in the future without using coroutines, since generating the nav mesh requires the collisions and active meshes to be updated for it to generate the latest nav mesh that adjusts to the current scene. Destroy() calls do not update until the next frame, so that would generate a nav mesh with the old collisions where the walls still existed, even if dungeon generation was done on Awake() and nav mesh generation on Start(), because destroy calls require 1 frame to update, forcing us to use solutions such as using coroutines with a delay to generate the nav mesh. This change fixes that potential issue.