Implemented ConnectRooms() method
The connect rooms will generate paths with a 1 tile width that go through the center points of the generate rooms. Allows connecting the rooms in a valid manner.

NOTE : in the future, when pre-made rooms are added for things such as boss fights, a more complex algorithm such as A* should be implemented, so that it prevents passing through already generated tiles. That way, it becomes easier to trace valid paths that go to a pre-selected entry point where the door will be located.