Created play menu scene
This scene is an intermediate menu which has the purpose of displaying different options to play, one of them being a campaign and the other a procedurally generated world.

This should be changed in the future, but I don't have time for this as of now, so this patchy solution is good enough.

Maybe replace with some "continue" and "new game" buttons or "skip intro" or whatever, idk...