Added audio listener to player prefab
The camera is located at a far distance from the player's location, which makes it harder to locate the source of an audio in space relative to the palyer's position. For better echolocation, the audio listener has been moved to the player prefab.

Note that the player mesh component is the one that rotates, not the player prefab's root, which means that audio will be correctly heard relative to the direction in which the camera is facing at all times, disregarding player rotation, which for this game's camera perspective would cause massive disorientation.