Modified logic in SoundManager to support using audio mixer
The audio mixer has a logarithmic scale in db, which means that a mapping function is required to map the linear 0-1 scale to this logarithmic scale that the audio mixer uses.

This new change also makes it easier to implement global sounds with specific sound types which play anywhere on the map, because now we don't need to handle specific pools of sounds on our own. We also no longer have the problem where an audio playing with one volume keeps playing with said volume even if we change the volume while it is playing in game (which could be manually fixed, but that would be a pain in the ass to deal with considering the time constraints we have...)