Added patch to fix issue with player control disabling
When disabling the player controls, the player would keep walking in the same direction if the controls were disabled while the walk input was held.

This has now been poorly patched away due to time constraints.

For future projects, it has become clear that we should have a specific update loop based input system, basically the classic. This whole event based system has only made things more complicated for the scale and deadlines of this project...