Created AudioSourcePoolController class
The AudioSourcePoolController class will be in charge of controlling and managing the pooling of a specific type of audio sources with an audio source prefab of the user's choosing.

This way, we will be able to standardise the audio playing system even further and simplifying the logic by moving a lot of the duplicated specific logic from the SoundManager to this class and making the SoundManager manager sound events through Audio source pools.