Moved ObjectPool from index based system to simpler iteration based system
The iteration based system consumes a little bit less memory but might have slower lookup loop times. It does seem to play better with Unity's object model tho, so it is better to use this and guarantee correct behaviour rather than trying to use broken optimizations.

TODO : Fix the index based system for faster lookup times in the future, figure out what the fuck is wrong with the other code...